plugin/pyproject.toml
[project]
name = "mega-code"
version = "1.0.0b1"
description = "Open-source Claude Code plugin — collect interaction data, extract skills, and optimize AI workflows"
readme = "README.md"
requires-python = ">=3.10,<3.14"
license = "Apache-2.0"
authors = [
    { name = "WisdomGraph", email = "support@megacode.ai" }
]
keywords = ["claude-code", "plugin", "skills", "optimization", "data-collection"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Software Development :: Libraries",
]
dependencies = [
    "httpx>=0.27.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
parquet = ["pyarrow>=12.0"]
telemetry = ["opentelemetry-api>=1.20.0", "opentelemetry-sdk>=1.20.0"]

[project.urls]
Homepage = "https://www.megacode.ai"
Repository = "https://github.com/wisdomgraph/mega-code"

[project.scripts]
mega-code = "mega_code.client.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["mega_code"]

[tool.hatch.build.targets.sdist]
include = [
    "mega_code/",
    ".claude-plugin/",
    "hooks/",
    "skills/",
    "scripts/",
    "README.md",
    "pyproject.toml",
]

[tool.black]
line-length = 100
target-version = ["py310"]

[tool.flake8]
max-line-length = 100
extend-ignore = ["E203", "W503"]

[tool.pytest.ini_options]
testpaths = ["tests"]
